# backend/application/src/main/resources/api/v1/specs/openapi.yaml
openapi: 3.0.3
info:
  title: AI Onboarding Platform API
  version: 1.0.0
  description: Backend API for AI Onboarding & Training platform
servers:
  - url: /api/v1
    description: V1 base path (documentation only)
...
paths:
...
  /api/v1/lessons/{id}/ask:
    post:
      summary: Ask lesson assistant
      description: Asks the lesson assistant a question with optional chat history.
      tags: [ lessons ]
      operationId: askLessonAssistant
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        description: Request payload for the askLessonAssistant.
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AskLessonRequestV1' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/AskLessonResponseV1' } } } }
        '401': { $ref: '#/components/responses/UnauthorizedV1' }
        '403': { $ref: '#/components/responses/ForbiddenV1' }
        '404': { $ref: '#/components/responses/NotFoundV1' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedV1:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorV1' }
    ForbiddenV1:
      description: Authenticated but lacking permission
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorV1' }
    NotFoundV1:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiErrorV1' }
...
  schemas:
...
    ChatRoleV1:
      description: Chat message author role.
      type: string
      enum: [ user, assistant ]
    LessonAssistantPresetV1:
      description: Assistant response mode requested by the learner. Missing preset means regular.
      type: string
      enum: [ regular, small_portions ]
...
    ApiErrorV1:
      description: Standard API error payload.
      type: object
      additionalProperties: false
      required: [ code, message, timestamp ]
      properties:
        code:
          description: Machine-readable code value.
          type: string
        message:
          description: Human-readable message text.
          type: string
        timestamp:
          description: Timestamp for the event or error.
          type: string
          format: date-time
        correlationId:
          description: Correlation identifier for tracing.
          type: string
        parameters:
          description: Parameters for Api Error.
          type: array
          items: { $ref: '#/components/schemas/ValidationParameterV1' }
    ValidationParameterV1:
      description: Single validation parameter attached to an API error.
      type: object
      additionalProperties: false
      required: [ code, value ]
      properties:
        code:
          description: Machine-readable code value.
          type: string
        value:
          description: Parameter value.
          type: string
...
    JsonMetadataV1:
      description: Open-ended JSON metadata map for provider or generation details.
      type: object
      additionalProperties: true
...
    ChatMessageV1:
      description: Chat Message payload.
      type: object
      additionalProperties: false
      required: [ role, content ]
      properties:
        role:
          description: Chat message author role.
          allOf:
            - $ref: '#/components/schemas/ChatRoleV1'
        content:
          description: Content for Chat Message.
          type: string
    AskLessonRequestV1:
      description: Ask Lesson Request payload.
      type: object
      additionalProperties: false
      required: [ question ]
      properties:
        question:
          description: Question text.
          type: string
        history:
          description: History for Ask Lesson Request.
          type: array
          items: { $ref: '#/components/schemas/ChatMessageV1' }
        preset:
          description: Assistant response mode. Missing preset means regular.
          allOf:
            - $ref: '#/components/schemas/LessonAssistantPresetV1'
    AskLessonResponseV1:
      description: Ask Lesson Response payload.
      type: object
      additionalProperties: false
      required: [ answer, metadata ]
      properties:
        answer:
          description: Assistant answer text.
          type: string
        metadata:
          description: Additional metadata map.
          allOf:
            - $ref: '#/components/schemas/JsonMetadataV1'
...